Tweak: Changed Output Directory
Tweak: Changed the output directory of the build file. Right now it caused too much problems regarding the building of the ISO and running the correlated profiledef.sh.

Besides that a few tweaks have been made to the packages file, it seems that I have used some packages that are out of date or not-exisiting anymore. I will need to dive in the docs again.